Fixed bounds function ignoring pivots with only one child
When acting upon a container of the following layout
 p1
 /
p2
where only 1 of the 2 containers for the pivot is sharded,
The bounds function, when called on p1, will use p2 as
the pivot for determining bounds. This will cause gt.p1
misplaced_objects query to include all objects before
and all objects after its pivot, thus dropping all objects
in the container.